# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["users"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["users"]}},"/api/users/{id}":{"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"]}},"/api/login":{"post":{"operationId":"LoginController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["login"]}},"/api/menu":{"post":{"operationId":"MenuController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMenuDto"}}}},"responses":{"201":{"description":""}},"tags":["menu"]},"get":{"operationId":"MenuController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["menu"]}},"/api/menu/{id}":{"get":{"operationId":"MenuController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["menu"]},"patch":{"operationId":"MenuController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMenuDto"}}}},"responses":{"200":{"description":""}},"tags":["menu"]},"delete":{"operationId":"MenuController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["menu"]}},"/api/roles":{"post":{"operationId":"RolesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":""}},"tags":["roles"]},"get":{"operationId":"RolesController_filterQuery","parameters":[],"responses":{"200":{"description":""}},"tags":["roles"]}},"/api/roles/all":{"get":{"operationId":"RolesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["roles"]}},"/api/roles/{id}":{"patch":{"operationId":"RolesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":""}},"tags":["roles"]}},"/api/roles/{id}/{name}":{"delete":{"operationId":"RolesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["roles"]}},"/api/upload":{"post":{"operationId":"UploadController_uploadFile","parameters":[],"responses":{"201":{"description":""}},"tags":["upload"]}}},"info":{"title":"JinPiKa","description":"nest-admin后台接口","version":"1.0","contact":{"name":"JinPiKa","url":"https://gitee.com/jinxin0517/nest-admin","email":"847164495@qq.com"}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"captcha":{"type":"string"}},"required":["captcha"]},"UpdateUserDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{"userName":{"type":"string","uniqueItems":true},"password":{"type":"string"},"loginType":{"type":"string","enum":["phone","account"]}},"required":["userName","password","loginType"]},"CreateMenuDto":{"type":"object","properties":{}},"CreateRoleDto":{"type":"object","properties":{}},"UpdateRoleDto":{"type":"object","properties":{}}}}}